+++ src/main/java/BlackjackApplication.java
@@ -1,4 +1,4 @@
-import controller.Game;
+import domain.controller.Game;
 import java.util.Scanner;
 import view.InputView;
 import view.OutputView;
