+++ src/test/java/domain/participant/PlayerTest.java
@@ -3,7 +3,6 @@ package domain.participant;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 
-import domain.Money;
 import domain.card.Card;
 import domain.card.CardDeck;
 import domain.card.TrumpNumber;
