+++ src/test/java/domain/participant/PlayersTest.java
@@ -4,7 +4,6 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.Money;
 import domain.card.Card;
 import domain.card.CardDeck;
 import domain.card.TrumpNumber;
